+++ PyPowerStore/tests/unit_tests/entity/file_system.py
@@ -1,6 +1,6 @@
 from PyPowerStore.utils import constants
 from PyPowerStore.tests.unit_tests.entity.base_abstract import Entity
-from PyPowerStore.tests.unit_tests.common_data import CommonData
+from PyPowerStore.tests.unit_tests.data.common_data import CommonData
 
 
 class FileSystemResponse(Entity):